linux_requirements.txt
anyio==3.5.0
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.2.2
camelot-py==0.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==35.0.0
cycler==0.10.0
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
google-api-core==1.31.3
google-auth==1.35.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
h11==0.12.0
html5lib==1.1
idna==2.10
Jinja2==3.0.3
kiwisolver==1.3.2
libcst==0.3.21
MarkupSafe==2.0.1
matplotlib==3.3.2
mypy-extensions==0.4.3
numpy==1.21.3
opencv-python==4.4.0.40
openpyxl==3.0.9
packaging==21.0
pandas==1.1.3
pdfminer.six==20211012
pdftotext==2.1.5
Pillow==8.2.0
proto-plus==1.19.6
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==3.0.1
PyPDF2==1.26.0
python-dateutil==2.8.2
python-Levenshtein==0.12.0
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.24.0
rich==10.12.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
starlette==0.17.1
telegram==0.0.1
tornado==6.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
tzdata==2021.5
tzlocal==4.0.2
urllib3==1.25.11
uvicorn==0.17.0
webencodings==0.5.1
